bpf: Add support for svc v2 to datapath
This commit introduces svc v2 support to the BPF datapath (to both -
lb and lxc).

First, this makes the svc lookup to rely only on services of
the v2 type. This is safe, as before reloading the datapath,
all v2 services will be created from legacy ones by the daemon.

Second, as we need to support upgrading/downgrading of cilium-agent
without disrupting open connections, both legacy and v2 service
endpoints are stored in CT entries.

Signed-off-by: Martynas Pumputis <m@lambda.lt>